app/pyproject.toml
#[tool.poetry]
#name = "myelectricaldata-import"
#version = "0.9.0"
#description = ""
#authors = ["m4dm4rtig4n <m4dm4rtig4n@gmail.com>"]
#readme = "README.md"
#packages = [{include = "myelectricaldata_import"}]
#
#[tool.poetry.dependencies]
#python = "^3.10"
#fastapi = "^0.100.0"
#paho-mqtt = "^1.6.1"
#python-dateutil = "^2.8.2"
#requests = "^2.31.0"
#Markdown = "^3.4.3"
#influxdb-client = "^1.36.1"
#PyYAML = "^6.0"
#mergedeep = "^1.3.4"
#Jinja2 = "^3.1.2"
#art = "^6.0"
#SQLAlchemy = "^2.0.18"
#alembic = "^1.11.1"
#waitress = "^2.1.2"
#rauth = "^0.7.3"
#pymysql = "^1.1.0"
#uvicorn = "^0.22.0"
#watchfiles = "^0.19.0"
#pytz = "^2023.3"
#
#
#[build-system]
#requires = ["poetry-core"]
#build-backend = "poetry.core.masonry.api"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "myelectricaldata_import"
version = "1"
dependencies = [
    "paho-mqtt",
    "python-dateutil",
    "requests",
    "markdown",
    "influxdb-client",
    "PyYAML",
    "mergedeep",
    "Flask",
    "Flask-APScheduler",
    "flask-swagger-ui",
    "jinja2",
    "art",
    "SQLAlchemy",
    "alembic",
    "waitress",
    "rauth",
    "psycopg2-binary",
    "PyMySQL",
    "fastapi",
    "uvicorn",
    "fastapi-utils",
]

[project.optional-dependencies]
dev = ["pytest"]